requirements.txt
Flask==2.2.2
Flask_MySQLdb==2.0.0
matplotlib==3.7.1
mysql_connector_repackaged==0.3.1
pandas==1.5.3
mysql-connector==2.2.9
